fix: properly access configuration keymaps
Problem:
When the user does not set keymaps in their
configuration, "config.keymaps" is nil, resulting
in an error in the call to ipairs.

Why:
ipairs receives "config.keymaps", but "M.config"
is the variable used to hold the tbl which is the
combination of the default config and the user
config.

Solution:
Pass M.config.keymaps to ipairs, not
config.keymaps.