fix: pasting
Problem: Pasting used to trigger a Lua error, because it was attempting
to call `vim.api,nvim_replace_termcodes` with `nil` as the first
parameter (`org_mapping.rhs`), and the function expected a valid
`string`.

This happened because the code was referring to `org_mapping`, which was
known to be an empty table in this conditional branch.

Moreover, the `keys` variable was not used, and the call to
`vim.api.nvim_feedkeys` referred to the original, unreplaced `opts.rhs`.

Solution: Use the correct `opts.rhs` argument when calling
`vim.api.nvim_replace_termcodes`, and then use the `keys` variable when
calling `vim.api.nvim_feedkeys`.

This fixes a bug that prevented the `p` and `P` keymaps from working
when this plugin was hijacking those mappings.